# internal-modules/ai-council/docs/custom-modfiles/openapi.json
{"openapi":"3.1.0","info":{"title":"Membership Initiation System","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/create-invitation":{"post":{"summary":"Create Invitation","operationId":"create_invitation_admin_create_invitation_post","parameters":[{"name":"invited_name","in":"query","required":true,"schema":{"type":"string","title":"Invited Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/validate-invitation":{"post":{"summary":"Validate Invitation","operationId":"validate_invitation_validate_invitation_post","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string","title":"Code"}},{"name":"pin","in":"query","required":true,"schema":{"type":"string","title":"Pin"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/submit-onboarding":{"post":{"summary":"Submit Onboarding","operationId":"submit_onboarding_submit_onboarding_post","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string","title":"Code"}},{"name":"voice_consent","in":"query","required":true,"schema":{"type":"boolean","title":"Voice Consent"}},{"name":"responses","in":"query","required":true,"schema":{"type":"string","title":"Responses"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/approve-membership":{"post":{"summary":"Approve Membership","operationId":"approve_membership_admin_approve_membership_post","parameters":[{"name":"invitation_code","in":"query","required":true,"schema":{"type":"string","title":"Invitation Code"}},{"name":"user_name","in":"query","required":true,"schema":{"type":"string","title":"User Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/validate-key":{"post":{"summary":"Validate Key","operationId":"validate_key_validate_key_post","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string","title":"Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}